split search to two component
