+++ src/app/app.component.ts
@@ -9,6 +9,7 @@ import { AuthUser } from './core/authUser';
 import { interval, observable, timer } from 'rxjs';
 import { switchMap } from 'rxjs/operators';
 import { Router, ActivatedRoute } from '../../node_modules/@angular/router';
+import { PlatformLocation } from '../../node_modules/@angular/common';
 
 @Component({
     selector: 'app-root',
@@ -21,42 +22,24 @@ export class AppComponent extends BaseComponent implements OnInit {
 
     constructor(
         protected logger: Logger,
-        private router: Router,
+        public router: Router,
         public msalService: MsalService,
         public msGraphService: MsGraphService
-    ) {
-        super(logger);
         
+    ) {
+        super(logger, router);
+
         this.msalService.authenticated.then((isAuthenticated: boolean) => {
             this.logger.info("isauth", isAuthenticated);
             if (!isAuthenticated) {
                 this.msalService.login();
             }
         })
+
     }
 
     ngOnInit() {
-    
+ 
     }
 
-    logout() {
-        this.msalService.logout();
-    }
-
-    testGraphApi() {
-        this.msalService.authenticated.then((isAuthenticated: boolean) => {
-            this.logger.info("isauth", isAuthenticated);
-        })
-
-        if (this.user.email) {
-            this.user = new AuthUser(this.user.name, this.user.email);
-            this.msGraphService.getPhotoByUpn(this.user.email).subscribe((photoBlob) => {
-                this.createImageFromBlob(photoBlob, this.user);
-            })
-
-            this.msGraphService.getUserProfile("danipi@M365x342201.onmicrosoft.com").subscribe((userProfile) => {
-                this.logger.info("profile", userProfile);
-            })
-        }
-    }
 }
