+++ src/app/component/home/home.component.html
@@ -23,7 +23,10 @@
             <div class="col col-lg-5 text-center">
                 <h1 class="text-center text-secondary mb-4">
                     <!-- <img class="mr-4" height="80" src="../../../assets/images/Microsoft_logo.png" alt="">  -->
-                    <span class="text-green">Help</span>&nbsp;<span class="text-yellow">Me</span>&nbsp;<span class="text-blue">Connect</span>
+                    <span class="text-red">H</span><span class="text-green">e</span><span class="text-blue">l</span><span class="text-yellow">p</span>&nbsp;
+                    <span class="text-red">M</span><span class="text-green">e</span>&nbsp;
+                    <span class="text-blue">C</span><span class="text-yellow">o</span><span class="text-red">n</span><span class="text-green">n</span>
+                    <span class="text-blue">e</span><span class="text-yellow">c</span><span class="text-red">t</span>
                     <!-- <img class="mr-4" height="0" src="../../../assets/images/user.png" alt=""> -->
                 </h1>
