+++ src/app/component/search/graph/graph.module.ts
@@ -1,3 +1,4 @@
+import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
@@ -5,24 +6,26 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { NgxGraphModule } from '@swimlane/ngx-graph';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
-import { SearchComponent } from './search.component';
-import { SearchService } from './search.service';
-import { SearchRoutingModule } from './search-routing.module';
+import { GraphRoutingModule } from './graph-routing.module';
+import { GraphComponent } from './graph.component';
+import { GraphService } from './graph.service';
+import { SearchHeaderModule } from '../../../templates/searchHeader/searchHeader.module';
 
 @NgModule({
     declarations: [
-        SearchComponent
+        GraphComponent
     ],
     imports: [
         CommonModule,
         HttpClientModule,
         FormsModule,
+        SearchHeaderModule,
+        GraphRoutingModule,
         NgxGraphModule,
         NgxChartsModule,
-        SearchRoutingModule,
     ],
     providers: [
-        SearchService
+        GraphService
     ]
 })
-export class SearchModule { }
+export class GraphModule { }
