--- src/app/component/search/search.component.css
@@ -1,60 +1,0 @@
-.user-photo {
-    width: 35px !important;
-}
-
-.logo{
-    height: 24px;
-}
-
-.input-group{
-    width: 600px;
-    position: relative;
-}
-
-.icon-search{
-    position: absolute;
-    right: 10px;
-    top: 15px;
-    z-index: 9
-}
-
-.search{
-    width: 600px;
-    height: 50px;
-    margin: 0 auto;
-    border-radius: 0;
-    outline: none !important;
-    box-shadow: none !important;
-}
-
-.pd-15{
-    padding: 15px;
-}
-
-.no-border{
-    border: none;
-}
-
-.box-shadow{
-    box-shadow: 0 0 15px rgba(0,0,0,0.5)
-}
-
-.hover:hover{
-    background: rgba(0,123,255,0.1)
-}
-
-.photo{
-    width: 50%;
-}
-
-.word-break-all {
-    word-break: break-all;
-}
-
-.w-80 {
-    width: 80%;
-}
-
-.pl-155{
-    padding-left: 155px
-}
