+++ src/app/component/search/search.component.html
@@ -1,85 +1,3 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light pt-4 mb-3">
-    <a class="navbar-brand text-secondary font-weight-bold d-inline" href="#">
-        <img class="mr-2 mb-1 logo" src="../assets/images/Microsoft_logo.png" />Microsoft
-    </a>
-
-    <div class="input-group">
-        <input class="form-control search" type="text" placeholder="Search" [(ngModel)]="query" (keyup.enter)="searchKeyWords(query)">
-        <div class="input-group-btn icon-search text-secondary">
-            <i class="fa fa-search fa-lg"></i>
-        </div>
-    </div>
-
-    <div class="collapse navbar-collapse" id="navbarNavDropdown">
-        <ul class="navbar-nav ml-auto">
-            <li class="nav-item dropdown mr-auto">
-                <a class="nav-link font-weight-bold d-inline" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true"
-                    aria-expanded="false" [innerHTML]="user.name">
-                </a>
-                <img class="rounded-circle ml-2 user-photo" *ngIf="user.photo" [src]="user.photo">
-                <img class="user-photo ml-2" *ngIf="!user.photo" src="../assets/images/user.png">
-                <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                    <a class="dropdown-item" href="#" (click)="logout()">Logout</a>
-                </div>
-            </li>
-        </ul>
-    </div>
-
-</nav>
-
-<div class="container-fluid pl-155">
-    <div class="text-center mt-4" *ngIf="isSearching && !isNotFound">
-        <i class="fa fa-spinner fa-pulse fa-3x fa-fw"></i>
-        <span class="sr-only">Loading...</span>
-    </div>
-    
-    <div class="alert alert-warning text-center" role="alert" *ngIf="isNotFound">
-        <span [innerHTML]="notFoundTip"></span>
-    </div>
-        
-    <div id="entity" *ngIf="!isSearching && !isNotFound">
-        <div class="row mb-5">
-            <div class="col-md-auto">
-                <h6 class="mb-3" *ngIf="entityCards && entityCards.length != 0">{{entityCards.length}} Results</h6>
-                <h3 *ngIf="entityCards && entityCards.length != 0">Top Ranked Experts who might know about "{{preQuery}}"</h3>
-            </div>
-        </div>
-        <div class="row">
-            <div class="col-lg-2" *ngFor="let entityCard of entityCards; trackBy: index">
-                <div class="card no-border hover box-shadow pd-15">
-                    <div class="text-center">
-                        <img class="rounded-circle pointer photo" *ngIf="entityCard.photo" [src]="entityCard.photo" alt="Card image cap" (click)="searchGraph(entityCard.email)">
-                    </div>
-                    <div class="card-body">
-                        <h5 class="card-title text-center" [innerHTML]="entityCard.name"></h5>
-                        <div class="text-center">
-                            <i class="fa fa-link mr-2"></i>
-                            <span class="font-weight-bold" [innerHTML]="entityCard.connection"></span>
-                        </div>
-                    </div>
-                    
-                    <ul class="list-unstyled">
-                        <li class="text-center">
-                            <!-- <i class="fa fa-briefcase mr-2"></i>
-                            <i class="fa fa-circle-o-notch fa-spin"></i> -->
-                            <span class="font-weight-bold" [innerHTML]="entityCard.jobTitle"></span>
-                        </li>
-                        <li class="text-center">
-                            <!-- <i class="fa fa-envelope mr-2"></i> -->
-                            <a class="font-weight-bold word-break-all" [href]="getEmailLink(entityCard.email)" [innerHTML]="entityCard.email"></a>
-                        </li>
-                        <li class="text-center">
-                            <!-- <i class="fa fa-phone mr-2"></i> -->
-                            <span class="font-weight-bold" [innerHTML]="entityCard.phone"></span>
-                        </li>
-                    </ul>
-
-                    <div class="text-center mt-4">
-                        <button type="button" class="btn btn-outline-primary">Details</button>
-                    </div>
-                    
-                </div>
-            </div>
-        </div>
-    </div>
+<div>
+    <router-outlet></router-outlet>
 </div>
