+++ src/app/component/search/topic/topic-routing.module.ts
@@ -1,23 +1,21 @@
 import { Routes, RouterModule } from "@angular/router";
 import { NgModule } from "@angular/core";
+import { TopicComponent } from "./topic.component";
 
-const searchRoutes: Routes = [
+const topicRoutes: Routes = [
     {
-        path: 'graph',
-        loadChildren: './graph/graph.module#GraphModule'
-    },
-    {
-        path: 'topic',
-        loadChildren: './topic/topic.module#TopicModule'
+        path: '',
+        component: TopicComponent
     }
 ];
 
+
 @NgModule({
     imports: [RouterModule.forChild(
-        searchRoutes,
+        topicRoutes,
         //his outputs each router event that took place during each navigation lifecycle to the browser console
         //{ enableTracing: true } // <-- debugging purposes only
     )],
     exports: [RouterModule]
 })
-export class SearchRoutingModule { }
+export class TopicRoutingModule { }
