+++ src/app/component/search/topic/topic.component.html
@@ -1,0 +1,58 @@
+<search-header (emitSearch)="searchKeyWords($event)"></search-header>
+
+<div class="container-fluid pl-155">
+    <div class="text-center mt-4" *ngIf="isSearching && !isNotFound">
+        <i class="fa fa-spinner fa-pulse fa-3x fa-fw"></i>
+        <span class="sr-only">Loading...</span>
+    </div>
+
+    <div class="alert alert-warning text-center" role="alert" *ngIf="isNotFound">
+        <span [innerHTML]="notFoundTip"></span>
+    </div>
+
+    <div id="entity" *ngIf="!isSearching && !isNotFound">
+        <div class="row mb-5">
+            <div class="col-md-auto">
+                <h6 class="mb-3" *ngIf="entityCards && entityCards.length != 0">{{entityCards.length}} Results</h6>
+                <h3 *ngIf="entityCards && entityCards.length != 0">Top Ranked Experts who might know about "{{preQuery}}"</h3>
+            </div>
+        </div>
+        <div class="row">
+            <div class="col-lg-2" *ngFor="let entityCard of entityCards; trackBy: index">
+                <div class="card no-border hover box-shadow pd-15">
+                    <div class="text-center">
+                        <img class="rounded-circle pointer photo" *ngIf="entityCard.photo" [src]="entityCard.photo" alt="Card image cap" (click)="search(entityCard.email)">
+                    </div>
+                    <div class="card-body">
+                        <h5 class="card-title text-center" [innerHTML]="entityCard.name"></h5>
+                        <div class="text-center">
+                            <i class="fa fa-link mr-2"></i>
+                            <span class="font-weight-bold" [innerHTML]="entityCard.connection"></span>
+                        </div>
+                    </div>
+
+                    <ul class="list-unstyled">
+                        <li class="text-center">
+                            <!-- <i class="fa fa-briefcase mr-2"></i>
+                            <i class="fa fa-circle-o-notch fa-spin"></i> -->
+                            <span class="font-weight-bold" [innerHTML]="entityCard.jobTitle"></span>
+                        </li>
+                        <li class="text-center">
+                            <!-- <i class="fa fa-envelope mr-2"></i> -->
+                            <a class="font-weight-bold word-break-all" [href]="getEmailLink(entityCard.email)" [innerHTML]="entityCard.email"></a>
+                        </li>
+                        <li class="text-center">
+                            <!-- <i class="fa fa-phone mr-2"></i> -->
+                            <span class="font-weight-bold" [innerHTML]="entityCard.phone"></span>
+                        </li>
+                    </ul>
+
+                    <div class="text-center mt-4">
+                        <button type="button" class="btn btn-outline-primary">Details</button>
+                    </div>
+
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
