+++ src/app/component/search/topic/topic.module.ts
@@ -6,24 +6,26 @@ import { FormsModule } from '@angular/forms';
 import { HttpClientModule } from '@angular/common/http';
 import { NgxGraphModule } from '@swimlane/ngx-graph';
 import { NgxChartsModule } from '@swimlane/ngx-charts';
-import { GraphRoutingModule } from './graph-routing.module';
-import { GraphComponent } from './graph.component';
-import { GraphService } from './graph.service';
+import { TopicComponent } from './topic.component';
+import { TopicService } from './topic.service';
+import { TopicRoutingModule } from './topic-routing.module';
+import { SearchHeaderModule } from '../../../templates/searchHeader/searchHeader.module';
 
 @NgModule({
     declarations: [
-        GraphComponent
+        TopicComponent
     ],
     imports: [
         CommonModule,
         HttpClientModule,
         FormsModule,
-        GraphRoutingModule,
+        SearchHeaderModule,
+        TopicRoutingModule,
         NgxGraphModule,
         NgxChartsModule,
     ],
     providers: [
-        GraphService
+        TopicService
     ]
 })
-export class GraphModule { }
+export class TopicModule { }
