+++ src/app/templates/searchHeader/searchHeader.component.css
@@ -1,3 +1,7 @@
+input[type=text]::-ms-clear{
+    display: none;
+}
+
 .user-photo {
     width: 35px !important;
 }
@@ -25,20 +29,4 @@
     border-radius: 0;
     outline: none !important;
     box-shadow: none !important;
-}
-
-.detail-infos{
-    padding: 15px;
-    background: #dcdcdc;
-    border: 1px solid #dddddd;
-}
-
-.strength-line{
-    display: inline-block;
-    width: 50px;
-    height: 5px;
-}
-
-.word-break-all {
-    word-break: break-all;
 }
