+++ src/app/templates/searchHeader/searchHeader.component.html
@@ -1,0 +1,28 @@
+<nav class="navbar navbar-expand-lg navbar-light bg-light pt-4 mb-3">
+    <a class="navbar-brand text-secondary font-weight-bold d-inline" href="#">
+        <img class="mr-2 mb-1 logo" src="../assets/images/Microsoft_logo.png" />Microsoft
+    </a>
+
+    <div class="input-group">
+        <input class="form-control search" type="text" placeholder="Search" [(ngModel)]="query" (keyup.enter)="emitSearchToRoot(query)">
+        <div class="input-group-btn icon-search text-secondary">
+            <i class="fa fa-search fa-lg"></i>
+        </div>
+    </div>
+
+    <div class="collapse navbar-collapse" id="navbarNavDropdown">
+        <ul class="navbar-nav ml-auto">
+            <li class="nav-item dropdown mr-auto">
+                <a class="nav-link font-weight-bold d-inline" href="#" id="navbarDropdownMenuLink" role="button" data-toggle="dropdown" aria-haspopup="true"
+                    aria-expanded="false" [innerHTML]="user.name">
+                </a>
+                <img class="rounded-circle ml-2 user-photo" *ngIf="user.photo" [src]="user.photo">
+                <img class="user-photo ml-2" *ngIf="!user.photo" src="../assets/images/user.png">
+                <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
+                    <a class="dropdown-item" href="#" (click)="logout()">Logout</a>
+                </div>
+            </li>
+        </ul>
+    </div>
+
+</nav>
