+++ src/app/templates/searchHeader/searchHeader.module.ts
@@ -1,0 +1,20 @@
+import { FormsModule } from "@angular/forms";
+import { NgModule } from "@angular/core";
+import { CommonModule } from "../../../../node_modules/@angular/common";
+import { SearchHeaderComponent } from "./searchHeader.component";
+
+
+@NgModule({
+    imports: [
+        CommonModule,
+        FormsModule
+    ],
+    exports: [
+        SearchHeaderComponent
+    ],
+    declarations: [
+        SearchHeaderComponent
+    ]
+})
+
+export class SearchHeaderModule { }
